refactor cli args in host
